move test controller preparation functions to testdoubles package to make it reusable & modify error handling
