+++ controller/tenant.go
@@ -106,7 +106,6 @@ func (c *TenantController) Clean(ctx *app.CleanTenantContext) error {
 		log.Error(ctx, params, "deletion of namespaces failed")
 		return jsonapi.JSONErrorResponse(ctx, err)
 	}
-
 	return ctx.NoContent()
 }
 
@@ -272,8 +271,15 @@ func (c *TenantController) Update(ctx *app.UpdateTenantContext) error {
 		return jsonapi.JSONErrorResponse(ctx, errors.NewNotFoundError("tenant", user.ID.String()))
 	}
 
-	TenantUpdater{Config: c.config, ClusterService: c.clusterService, TenantService: c.tenantService}.
+	err = TenantUpdater{Config: c.config, ClusterService: c.clusterService, TenantService: c.tenantService}.
 		Update(ctx, dbTenant, user, environment.DefaultEnvTypes, true)
+	if err != nil {
+		log.Error(ctx, map[string]interface{}{
+			"err":      err,
+			"tenantID": dbTenant.ID,
+		}, "update of namespaces failed")
+		return jsonapi.JSONErrorResponse(ctx, err)
+	}
 
 	ctx.ResponseData.Header().Set("Location", rest.AbsoluteURL(ctx.RequestData.Request, app.TenantHref()))
 	return ctx.Accepted()
@@ -290,17 +296,13 @@ func (u TenantUpdater) Update(ctx context.Context, dbTenant *tenant.Tenant, user
 	// get tenant's namespaces
 	namespaces, err := tenantRepository.GetNamespaces()
 	if err != nil {
-		log.Error(ctx, map[string]interface{}{
-			"err":      err,
-			"tenantID": dbTenant.ID,
-		}, "retrieval of existing namespaces from DB failed")
-		return jsonapi.JSONErrorResponse(ctx, err)
+		return errs.Wrap(err, "retrieval of existing namespaces from DB failed")
 	}
 
 	// create cluster mapping from existing namespaces
 	clusterMapping, err := GetClusterMapping(ctx, u.ClusterService, namespaces)
 	if err != nil {
-		return jsonapi.JSONErrorResponse(ctx, errors.NewInternalError(ctx, err))
+		return err
 	}
 
 	// create openshift service
@@ -321,24 +323,7 @@ func (u TenantUpdater) Update(ctx context.Context, dbTenant *tenant.Tenant, user
 	openShiftService := openshift.NewService(serviceContext, nsRepo, envService)
 
 	// perform patch method on the list of exiting namespaces
-	err = openShiftService.Update(envTypes, namespaces, openshift.UpdateOpts().EnableSelfHealing())
-	if err != nil {
-		log.Error(ctx, map[string]interface{}{
-			"err":                err,
-			"tenantID":           dbTenant.ID,
-			"namespacesToUpdate": listNames(namespaces),
-		}, "update of namespaces failed")
-		return jsonapi.JSONErrorResponse(ctx, err)
-	}
-	return nil
-}
-
-func listNames(namespaces []*tenant.Namespace) []string {
-	var names []string
-	for _, ns := range namespaces {
-		names = append(names, ns.Name)
-	}
-	return names
+	return openShiftService.Update(envTypes, namespaces, openshift.UpdateOpts().EnableSelfHealing())
 }
 
 func (c *TenantController) getExistingTenant(ctx context.Context, id uuid.UUID, osUsername string) (*tenant.Tenant, error) {
